fix: renamed files casing to match git defaults
